requirements.txt
python-dotenv
google-generativeai==0.7.0
imagehash
annotated-types==0.7.0
asgiref==3.9.1
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
decorator==5.2.1
Django==5.2.5
executing==2.2.0
google-ai-generativelanguage==0.6.5
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.62.3
httplib2==0.30.0
idna==3.10
ImageHash==4.3.2
ipdb==0.13.13
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.3.2
parso==0.8.5
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow-heif==0.16.0
python-magic==0.4.27
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.8
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dotenv==1.1.1
PyWavelets==1.9.0
requests==2.32.5
rsa==4.9.1
scipy==1.16.1
sqlparse==0.5.3
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13